feat(api): implement goal creation and integrate with session handling
- Added a new API handler for creating goals with validation using Zod.
- Updated session creation to allow for associated goals, ensuring proper session ownership.
- Modified database schema to change the result type of goals from numeric to integer.
- Enhanced error handling for goal creation and session management.

+++ packages/web/api/lib/db/schema/index.ts
@@ -95,7 +95,7 @@ export const goals = sqliteTable(
 			autoIncrement: true,
 		}),
 		key: text().notNull(),
-		result: numeric().notNull(),
+		result: integer().notNull(),
 		unit: text().notNull(),
 		description: text(),
 		weightage: integer().default(0),
